aggregation: renamed `AggregatePlot` to `PlotEach`
